# mechanic-shop-backend/app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop Management API"
  description: "Used to manage the everyday functionalities of a mechanic shop"
  version: "0.0.1"

host: "https://mechanic-shop-db-model-2.onrender.com" #During development my API is served on my local host (will change when we go to prod)
schemes:
  - "https"
consumes:
  - "application/json" #recieves json data
produces:
  - "application/json" #produces json data

securityDefinitions: #Describing how to add token auth to the endpoints
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header

paths: #Information about our routes

  /mechanics/login: #endpoint starts the path
    post: #request type (post,get,put,delete)
      tags:
        - Mechanics #Adds this route to a category
      summary: "Login endpoint (generates a token)" 
      description: "Validates mechanic credentials (email and password), and produces a user specific token." 
      parameters: 
        - in: "body" #Where will this info come from
          name: "body"
          description: "Login Credentials" #What are you sending
          required: true 
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login Successful"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              message: "Welcome Mechanic"
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOj..."

  /mechanics:
    post: #Create Mechanic Endpoint
      tags:
        - Mechanics
      summary: "Register Mechanic"
      description: "Send in mechanic info, to create a mechanic profile that will be stored in the database."
      parameters:
        - in: "body"
          name: "body"
          description: "New Mechanic Information"
          required: true
          schema:
            $ref: "#/definitions/MechanicInput"
          
      responses:
        201:
          description: "Creation Successful"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              first_name: "John"
              last_name: "Doe"
              address: "123 Fun St."
              email: "test@test1.com"
              id: 1
              password: "<encrypted password>"
              salary: 90500.50
    
    get: #Get all Mechanics Endpoint
      tags:
        - Mechanics
      summary: "Returns all Mechanics"
      description: "Endpoint to recieve an array of mechanic objects."

      responses:
        200:
          description: "Successfully Retrieved Mechanics"
          schema:
            $ref: "#/definitions/MechanicsResponse"

  /mechanics/{mechanic_id}:
    get: #View individual mechanic
      tags:
        - Mechanics
      summary: "Get a single mechanic"
      parameters:
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
          schema:
            $ref: "#/definitions/MechanicInput"

      responses:
        200:
          description: "Successfully Retreived Mechanic"
          schema:
            $ref: "#/definitions/MechanicResponse"

    put: #Update a Mechanic
      tags:
        - Mechanics
      summary: "Update Mechanic Endpoint"
      description: "Endpoint to update a mechanic, requires a token, mechanic will be identified by the token"
      security:
        - bearerAuth: [] #Referencing our Token Auth definition (Make sure when authorizing in Swagger you add Bearer <Token>)
      parameters:
        - in: "body"
          name: "body"
          description: "Mechanic Updates"
          required: true
          schema:
            $ref: "#/definitions/MechanicUpdate" 
          
      responses:
        200:
          description: "Update Successful"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              first_name: "John"
              last_name: "Doe"
              address: "123 Fun St."
              email: "test@test1.com"
              id: 1
              password: "<encrypted password>"
              salary: 90500.50
    
    delete: #Delete a mechanic
      tags:
        - Mechanics
      summary: "Delete a Mechanic"
      description: "Delete a mechanic, requires a token, token will be used to identify mechanic to delete."
      security:
        - bearerAuth: []

      responses:
        200:
          description: "Deleted Mechanic"
          schema:
            type: object
            properties:
              message:
                type: string

  /mechanics/my-tickets:
    get: 
      tags:
        - Mechanics
      summary: "Get tickets related to the authenticated mechanic"
      security:
        - bearerAuth: []

      responses:
        200:
          description: "Successfully Retreived Service Tickets of Mechanic"
          schema:
            $ref: "#/definitions/ServiceTicketResponse"

  /customers:
    post: #Create Customer Endpoint
      tags:
        - Customers
      summary: "Register Customer"
      description: "Send in customer info, to create a customer profile that will be stored in the database."
      parameters:
        - in: "body"
          name: "body"
          description: "New Customer Information"
          required: true
          schema:
            $ref: "#/definitions/CustomerInput"
          
      responses:
        201:
          description: "Creation Successful"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              first_name: "John"
              last_name: "Doe"
              address: "123 Fun St."
              email: "test@test.com"
              phone: "555-555-5555"
              id: 1
    
    get: #Get all Customers Endpoint
      tags:
        - Customers
      summary: "Returns all Customers"
      description: "Endpoint to recieve an array of customers objects."

      responses:
        200:
          description: "Successfully Retrieved Customers"
          schema:
            $ref: "#/definitions/CustomersResponse"

  /customers/{customer_id}:
    get: #View individual Customer
      tags:
        - Customers
      summary: "Get a single customer"
      parameters:
        - in: "path"
          name: "customer_id"
          type: "integer"
          required: true
          schema:
            $ref: "#/definitions/CustomerInput"

      responses:
        200:
          description: "Successfully Retreived Customer"
          schema:
            $ref: "#/definitions/CustomerResponse"

    put: #Update a Customer
      tags:
        - Customers
      summary: "Update Customer Endpoint"
      description: "Endpoint to update a customer"
      parameters:
        - in: "body"
          name: "body"
          description: "Customer Updates"
          required: true
          schema:
            $ref: "#/definitions/CustomerUpdate" 
          
      responses:
        200:
          description: "Update Successful"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              first_name: "John"
              last_name: "Doe"
              address: "123 Fun St."
              email: "test@test.com"
              phone: "555-555-5556"
              id: 1
    
    delete: #Delete a customer
      tags:
        - Customers
      summary: "Delete a Customer"
      description: "Endpoint to delete a customer"
      parameters:
        - in: "path"
          name: "customer_id"
          type: "integer"
          required: true

      responses:
        200:
          description: "Deleted Customer"
          schema:
            type: object
            properties:
              message:
                type: string

  /service_tickets:
    post: #Create Service Tickets Endpoint
      tags:
        - ServiceTickets
      summary: "Register Service Ticket"
      description: "Send in service ticket info, to create a service ticket profile that will be stored in the database."
      parameters:
        - in: "body"
          name: "body"
          description: "New Service Ticket Information"
          required: true
          schema:
            $ref: "#/definitions/ServiceTicketInput"
          
      responses:
        201:
          description: "Creation Successful"
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
          examples:
            application/json:
              customer_id: 1
              service_desc: "Oil change"
              price: 125.75
              vin: "usa219324b3289re9df3"
              service_date: "1900-01-01"
              id: 1

    get: #Get all Service Ticket Endpoint
      tags:
        - ServiceTickets
      summary: "Returns all Service Tickets"
      description: "Endpoint to recieve an array of service tickets objects."

      responses:
        200:
          description: "Successfully Retrieved Service Tickets"
          schema:
            $ref: "#/definitions/ServiceTicketsResponse"

  /service_tickets/{service_ticket_id}:  
    get: #Get individual Service Ticket Endpoint
      tags:
        - ServiceTickets
      summary: "Returns a Service Ticket"
      description: "Endpoint to recieve a service ticket object."
      parameters:
        - in: "path"
          name: "service_ticket_id"
          type: "integer"
          required: true
          schema:
            $ref: "#/definitions/ServiceTicketInput"

      responses:
        200:
          description: "Successfully Retrieved Service Ticket"
          schema:
            $ref: "#/definitions/ServiceTicketResponse"

    put: #Update a Service Ticket
      tags:
        - ServiceTickets
      summary: "Update Service Ticket Endpoint"
      description: "Endpoint to update a service ticket"
      parameters:
        - in: "body"
          name: "body"
          description: "Service Ticket Updates"
          required: true
          schema:
            $ref: "#/definitions/ServiceTicketUpdate" 
          
      responses:
        200:
          description: "Update Successful"
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
          examples:
            application/json:
              first_name: "John"
              last_name: "Doe"
              address: "123 Fun St."
              email: "test@test.com"
              phone: "555-555-5558"
              id: 1
    
    delete: #Delete a Service Ticket
      tags:
        - ServiceTickets
      summary: "Delete a Service Ticket"
      description: "Endpoint to delete a Service Ticket"
      parameters:
        - in: "path"
          name: "service_ticket_id"
          type: "integer"
          required: true

      responses:
        200:
          description: "Deleted Service Ticket"
          schema:
            type: object
            properties:
              message:
                type: string

  /service_tickets/{service_ticket_id}/add-mechanic/{mechanic_id}:
    put:
      tags:
        - TicketMechanics
      summary: "Add mechanic to Service Ticket Endpoint"
      description: "Endpoint to add a mechanic to a service ticket through the junction table, Ticket Mechanics"
      parameters:
        - in: "path"
          name: "service_ticket_id"
          type: "integer"
          required: true
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
          
      responses:
        200:
          description: "Mechanic Successfully Added"
          examples:
            application/json:
              mechanic_id: 1
              ticket_id: 1

  /service_tickets/{service_ticket_id}/remove-mechanic/{mechanic_id}:
    put:
      tags:
        - TicketMechanics
      summary: "Remove mechanic from Service Ticket Endpoint"
      description: "Endpoint to remove a mechanic from a service ticket through the junction table, Ticket Mechanics"
      parameters:
        - in: "path"
          name: "service_ticket_id"
          type: "integer"
          required: true
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
          
      responses:
        200:
          description: "Mechanic Successfully Removed"
          examples:
            application/json:
              mechanic_id: 1
              ticket_id: 1
              
  /service_tickets/{service_ticket_id}/add-part/{part_description_id}:
    put:
      tags:
        - ServiceTickets
      summary: "Add Part to Service Ticket Endpoint"
      description: "Endpoint to add a part to a service ticket through the junction table, Parts"
      parameters:
        - in: "path"
          name: "service_ticket_id"
          type: "integer"
          required: true
        - in: "path"
          name: "part_description_id"
          type: "integer"
          required: true 
          
      responses:
        200:
          description: "Part Successful Added"
          examples:
            application/json:
              desc_id: 1
              ticket_id: 1

  /parts/{description_id}:
    post: #Create Part Endpoint
      tags:
        - Parts
      summary: "Create Parts (one or more) from a Part Description id"
      description: "Creates one or more Part rows for the given Part Description, quantity set via query param 'qty' (default=1)."
      parameters:
        - in: "path"
          name: "description_id"
          type: "integer"
          required: true 
        - in: "body"
          name: "body"
          description: "New Part Information"
          required: true
          
      responses:
        201:
          description: "Creation Successful"
          examples:
            application/json:
              desc_id: 1
              ticket_id: 1
              id: 1

  /parts:  
    get: #Get all Parts Endpoint
      tags:
        - Parts
      summary: "Returns all Parts"
      description: "Endpoint to recieve an array of parts objects."
      
      responses:
        200:
          description: "Successfully Retrieved Parts"
          schema:
            $ref: "#/definitions/PartsResponse"

  /parts/{part_id}:  
    get: #Get a Part Endpoint
      tags:
        - Parts
      summary: "Return a single Part"
      description: "Endpoint to recieve a single part object."
      parameters:
        - in: "path"
          name: "part_id"
          type: "integer"
          required: true 

      responses:
        200:
          description: "Successfully Retrieved Part"
          schema:
            $ref: "#/definitions/PartResponse"

    delete: #Delete a Part
      tags:
        - Parts
      summary: "Delete a part"
      description: "Endpoint to delete a part"
      parameters:
        - in: "path"
          name: "part_id"
          type: "integer"
          required: true 

      responses:
        200:
          description: "Successfully Deleted Part"
          schema:
            type: object
            properties:
              message:
                type: string

  /parts/{part_id}/descriptions:
    get:
      tags:
        - Parts
      summary: "View Part Description for a Part"
      description: "Endpoint to view Part Description info for a part"
      parameters:
        - in: "path"
          name: "part_id"
          type: "integer"
          required: true

      responses:
        200:
          description: "Successfully Retreived Part Description"
          schema:
            $ref: "#/definitions/PartDescriptionResponse"

    put: #Update a Part Description
      tags:
        - PartDescriptions
      summary: "Update Part Description Endpoint"
      description: "Endpoint to update a Part Description"
      parameters:
        - in: "body"
          name: "body"
          description: "Part Description Updates"
          required: true
          schema:
            $ref: "#/definitions/PartDescriptionUpdate" 
          
      responses:
        200:
          description: "Creation Successful"
          schema:
            $ref: "#/definitions/PartDescriptionResponse"
          examples:
            application/json:
              part_name: "Honda oil filter"
              price: 24.99
              id: 1

  /parts/descriptions:
    post: #Create Part Description Endpoint
      tags:
        - PartDescriptions
      summary: "Create Part Description"
      description: "Send in part description info, to create a part description profile that will be stored in the database."
      parameters:
        - in: "body"
          name: "body"
          description: "New Part Description Information"
          required: true
          schema:
            $ref: "#/definitions/PartDescriptionInput"
          
      responses:
        201:
          description: "Creation Successful"
          schema:
            $ref: "#/definitions/PartDescriptionResponse"
          examples:
            application/json:
              part_name: "K&N oil filter"
              price: 24.99
              id: 1

    get:
      tags:
        - Parts
      summary: "View all Part Descriptions"
      description: "Endpoint to view all Part Descriptions"

      responses:
        200:
          description: "Successfully Retreived all Part Descriptions"
          schema:
            $ref: "#/definitions/PartDescriptionsResponse"

  /parts/descriptions/{part_description_id}:
    delete: #Delete a Part Description
      tags:
        - PartDescriptions
      summary: "Delete a Part Description"
      description: "Endpoint to delete a Part Description"
      parameters:
        - in: "path"
          name: "part_description_id"
          type: "integer"
          required: true

      responses:
        200:
          description: "Successful Deleted Part Description"
          schema:
            type: object
            properties:
              message:
                type: string



definitions: #Information about the shape of input and output json

  LoginCredentials: #Input data required for login route
    type: object
    properties:
      email:
        type: string
      password:
        type: string

    required:
      - email
      - password

  LoginResponse: #Response from /login endpoint
    type: object
    properties:
      message:
        type: string
      token:
        type: string

  MechanicInput: #Sending mechanic Information
    type: object
    properties:
      first_name:
        type: string
        example: "John"
      last_name:
        type: string
        example: "Doe"
      address:
        type: string
        example: "123 Fun St."
      email:
        type: string
        example: "test@test1.com"
      password:
        type: string
        example: "123"
      salary:
        type: float
        example: 90500.50

    required:
      - email
      - password
      - first_name
      - last_name
      - salary
      - address

  MechanicUpdate:
    type: object
    properties:
      first_name:
        type: string
      last_name:
        type: string
      address:
        type: string
      email:
        type: string
      password:
        type: string
      salary:
        type: float

  MechanicResponse: #Recieve Single Mechanic Profile
    type: object
    properties:
      id: 
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      address:
        type: string
      email:
        type: string
      password:
        type: string
      salary:
        type: float

  MechanicsResponse: #Response for multiple Mechanics 
    type: array
    items:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        address:
          type: string
        email:
          type: string
        password:
          type: string
        salary:
          type: float

  CustomerInput: #Sending customer Information
    type: object
    properties:
      first_name:
        type: string
        example: "John"
      last_name:
        type: string
        example: "Doe"
      address:
        type: string
        example: "123 Fun St."
      email:
        type: string
        example: "test@test.com"
      phone:
        type: string
        example: "555-555-5555"
      

    required:
      - email
      - first_name
      - last_name
      - phone
      - address

  CustomerUpdate: #Update Customer Profile
    type: object
    properties:
      first_name:
        type: string
      last_name:
        type: string
      address:
        type: string
      email:
        type: string
      phone:
        type: string

  CustomerResponse: #Recieve Single Customer Profile
    type: object
    properties:
      id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      address:
        type: string
      email:
        type: string
      phone:
        type: string

  CustomersResponse: #Response for multiple Customers 
    type: array
    items:
      type: object
      properties:
        id: 
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        address:
          type: string
        email:
          type: string
        phone:
          type: string
          
  ServiceTicketInput:
    type: object
    properties:
      customer_id:
        type: integer
        example: 1
      service_desc:
        type: string
        example: "Oil change"
      price:
        type: float
        example: 125.75
      vin:
        type: string
        example: "1HGCM82633A004352"
      service_date:
        type: string
        format: date-time


    required:
      - customer_id
      - service_desc
      - price
      - vin
      - service_date

  ServiceTicketUpdate:
    type: object
    properties:
      customer_id:
        type: integer
        example: 1
      service_desc:
        type: string
        example: "Oil change"
      price:
        type: float
        example: 125.75
      vin:
        type: string
        example: "1HGCM82633A004352"
      service_date:
        type: string
        format: date-time

  ServiceTicketResponse:
    type: object
    properties:
      id: 
        type: integer
        example: 1
      customer_id:
        type: integer
        example: 1
      service_desc:
        type: string
        example: "Oil change"
      price:
        type: float
        example: 125.75
      vin:
        type: string
        example: "1HGCM82633A004352"
      service_date:
        type: string
        format: date-time

  ServiceTicketsResponse:
    type: array
    items:
      id: 
        type: integer
        example: 1
      customer_id:
        type: integer
        example: 1
      service_desc:
        type: string
        example: "Oil change"
      price:
        type: float
        example: 125.75
      vin:
        type: string
        example: "1HGCM82633A004352"
      service_date:
        type: string
        format: date-time

  PartDescriptionInput:
    type: object
    properties:
      part_name:
        type: string
        example: "Brake Pad"
      price:
        type: float
        example: 50.0


    required:
      - part_name
      - price

  PartDescriptionUpdate:
    type: object
    properties:
      part_name:
        type: string
        example: "Brake Pad"
      price:
        type: float
        example: 50.0

  PartDescriptionResponse:
    type: object
    properties:
      id: 
        type: integer
      part_name:
        type: string
        example: "Brake Pad"
      price:
        type: number
        format: float
        example: 50.0

  PartDescriptionsResponse:
    type: array
    items:
      id: 
        type: integer
      part_name:
        type: string
        example: "Brake Pad"
      price:
        type: float
        example: 50.0

  PartResponse:
    type: object
    properties:
      id:       
        type: integer 
      desc_id:  
        type: integer 
      ticket_id:
        type: integer

  PartsResponse:
    type: array
    items:
      id:       
        type: integer 
      desc_id:  
        type: integer 
      ticket_id:
        type: integer